requirements.txt
beautifulsoup4==4.11.1
paho_mqtt==1.6.1
python_dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
